Add `/people/:id` routes and starter page
